notifications_system/requirements/local.txt
channels==4.1.0
channels-redis==4.2.0
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-redis==5.4.0
psycopg2-binary==2.9.9
pytest==8.3.2
#python-dotenv==1.0.1
requests==2.32.3

daphne==4.1.2
gunicorn==23.0.0



pdbp==1.5.3
